% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/env.fn.R
\name{env.fn}
\alias{env.fn}
\title{Evaluate the plateau envelope function}
\usage{
env.fn(pars, x.clim, slope.limit = 25)
}
\arguments{
\item{pars}{The vector of envelope parameters, length 2p+p+2+p(p-1)/2.}

\item{x.clim}{The n by p matrix of climate covariates.}

\item{slope.limit}{A scalar putting an upper bound on the envelope slopes;
limit is approximately \code{exp(slope.limit)}.}
}
\value{
A list of several named objects: the evaluations of the envelope
in a vector \code{x.envelope}; and two objects (\code{beta.mat} and
\code{ax.vec}) used by the \code{glm.env.fn} function when penalising
the deviance.
}
\description{
\code{env.fn} evaluates the plateau envelope for a given set of parameters
and a given set of climate covariates. Called by the internal fitting
function \code{glm.env.fn} and the map plotting function \code{plot.map}.
}

