% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/envelope.plot.R
\name{envelope.plot}
\alias{envelope.plot}
\title{Plot the plateau envelopes produced by \code{fit.glm.env} and
\code{fit.bugs.env}.}
\usage{
envelope.plot(envelope.fit, type = "persp", x.plot.lims = c(0, 1),
  len = 100, plot.vars, x.labels, fix.values, close.points = "",
  contour.levels)
}
\arguments{
\item{envelope.fit}{The fitted object from either \code{fit.glm.env} or
\code{fit.bugs.env}.}

\item{type}{String taking values \code{"persp"} or \code{"contour"}
to specify which type of plot is required; ignored if fitted object has
only one climate variable (when a univariate plot with rug is plotted).}

\item{x.plot.lims}{Vector of two numerics containing the endpoints for the
plotting regions in the (scaled) climate variables.}

\item{len}{The number of points to use for plotting in each dimension.}

\item{plot.vars}{If p>2 (currentl implying p==3), need to know which
two variables to plot; this vector of two climate column names makes
the selection.}

\item{x.labels}{Labels to plot on the x-axes, notionally the variable
names; if omitted, the code will use "Variable 1" etc.}

\item{fix.values}{Vector of numeric length p with named elements
corresponding to the values (and names) of the non-plotted variables;
ignored if p<3.}

\item{close.points}{String indicating how to plot points in contour or
perspective plots: the value "shade" plots filled circles for points
which are "close" to the chosen values of the held out variables, and
empty circles otherwise; the value "plot" is as for shade but without
the empty circles; and for any other value of \code{close.points},
the full data set will be plotted as empty circles.}

\item{contour.levels}{If specified, the probability values to be plotted
in the contour plots; this is especially important when using
\code{plot.envelope} to make a "video".}
}
\value{
For the case of a single climate variable, the plot produced is a
line graph, possibly with rugplot and points. For two climate covariates,
either a contour plot or a perspective plot is produced. For three climate
covariates, the same choice of contour or perspective plot is made for two
of the variables for a fixed value of the third covariate.
}
\description{
\code{envelope.plot} plots the plateau envelopes produced by the functions
\code{fit.glm.env} and \code{fit.bugs.env}.
}
\seealso{
\code{link{map.plot}}, \code{link{fit.bugs.env}}
}

