% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generate.initial.values.R
\name{generate.initial.values}
\alias{generate.initial.values}
\title{Generate initial values for plateau envelopes}
\usage{
generate.initial.values(y, x.clim, x.nonclim = NULL, constrain.beta = FALSE,
  random = FALSE, pars)
}
\arguments{
\item{y}{The binary response variable (taking values 0 or 1 for absence and
presence respectively).}

\item{x.clim}{The n by p matrix of climate covariates.}

\item{x.nonclim}{The n by p2 matrix of non-climate covariates.}

\item{constrain.beta}{Should ridge penalty be imposed on the betas (slopes)?}

\item{random}{Logical flag on whether or not to generate random starting values}

\item{pars}{The vector of envelope parameters, length 2p+p+2+p(p-1)/2}
}
\value{
If \code{random==TRUE}, the output will be a single vector of
parameter values in the order required for \code{glm.env.fn}; if
\code{random==FALSE}, the output will be a list with two elements:
\code{initial.pars}, a single vector of parameter values in the order
required for \code{glm.env.fn}, and \code{constrain.beta}, also as for
\code{glm.env.fn}.
}
\description{
\code{generate.initial.values} generates initial values for the plateau
envelopes. It is called by \code{fit.glm.env} and \code{fit.bugs.env}, and is
not intended for use interactively.
}

