% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{plateau-package}
\alias{plateau}
\alias{plateau-package}
\title{Fit plateau climate envelopes in species distribution models for presence/absence data}
\description{
Fit plateau climate envelopes in species distribution models for presence/absence data
}
\details{
\tabular{ll}{
Package: \tab plateau\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-12-18\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

\code{\link{plot.envelope}} plots an envelope
}
\examples{
plot.envelope()
}
\author{
Mark J Brewer \email{Mark.Brewer@bioss.ac.uk}
}
\references{
\url{http://www.bioss.ac.uk}
}
\keyword{package}

