% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{plateau}
\alias{plateau}
\alias{plateau-package}
\title{Fit plateau climate envelopes in species distribution models for presence/absence data}
\description{
plateau allows you to fit "plateau" (and unimodal) climate envelopes in
species distribution models for presence/absence data
}

