% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope.predict.R
\name{envelope.predict}
\alias{envelope.predict}
\title{Predict values for the plateau envelope models produced by \code{fit.glm.env} and
\code{fit.bugs.env}.}
\usage{
envelope.predict(envelope.fit, newdata)
}
\arguments{
\item{envelope.fit}{The fitted object from either \code{fit.glm.env} or
\code{fit.bugs.env}.}

\item{newdata}{A matrix or data frame of new data for prediction.}
}
\value{
A set of predictions from the supplied fitted model evaluated at the covariate values in
object \code{newdata}.
}
\description{
\code{envelope.predict} predicts values for the plateau envelope models produced by the functions
\code{fit.glm.env} and \code{fit.bugs.env}.
}
\seealso{
\code{link{map.plot}}, \code{link{fit.bugs.env}}
}

