% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fit.glm.env.R
\name{fit.glm.env}
\alias{fit.glm.env}
\title{Fit a plateau envelope via a (non-spatial) GLM and non-linear optimisation}
\usage{
fit.glm.env(y, x.clim, x.nonclim = NULL, initial.pars.input,
  random.search = FALSE, n.iter = 100, constrain.beta = FALSE,
  slope.limit = 7)
}
\arguments{
\item{y}{The binary response variable (taking values 0 or 1 for absence and
presence respectively).}

\item{x.clim}{The n by p matrix of climate covariates.}

\item{x.nonclim}{The n by p2 matrix of non-climate covariates.}

\item{initial.pars.input}{A vector of length 2p+p+2+p(p-1)/2 containing
starting values for each parameter; if missing, the code will suggest
starting values.}

\item{random.search}{A logical variable, which when set to \code{TRUE}
will cause the generation of \code{n.iter} random sets of starting values.}

\item{n.iter}{If \code{random.search} is \code{TRUE}, then \code{n.iter}
is number of sets of random starting values to use.}

\item{constrain.beta}{A p by 2 matrix of logicals in order to indicate
which beta parameters should be constrained to not vary too
much from its pair - set at most only one of these to \code{TRUE}
in each row. A value of \code{TRUE} in column 1 suggests that
there are not enough data to estimate the "up" (left-hand)
part of the envelope, and in column 2 correspondingly for
the "down" (right-hand) part of the envelope. If
betas should be constrained.}

\item{slope.limit}{Scalar putting an upper bound on the envelope slopes;
limit is approximately \code{exp(slope.limit)}.}
}
\value{
A list containing the following elements:
\describe{
\item{\code{par}}{Vector of estimates of the parameters in the order: (2*p) *log*
                   of slopes beta_11,beta_12,beta21,beta22,...; (p) apex
                   coordinates of the climate variables ax; (1) *log* distance
                   beneath the y apex of the "slice" parameter; (1) the apex
                   coordinate in the y direction; and if p>1, (p*(p-1)/2)
                   pairwise interaction terms.}
\item{\code{value}}{The (scalar) value of the "optimised" parameters.}
\item{\code{counts}}{The number of \code{glm.env.fn} function calls
(\code{NA} for gradient).}
\item{\code{convergence}}{Report from \code{optim()} as to convergence
status; beware that 0 does not reliably mean "good" convergence.}
\item{\code{message}}{Any message from \code{optim()}.}
\item{\code{y}}{The response data, returned mainly to aid the plotting
functions.}
\item{\code{x.clim}}{The climate data, also for the plotting functions.}
}
}
\description{
\code{fit.glm.env} fits a plateau envelope via a (non-spatial) GLM and
non-linear optimisation. This function exists to provide scope for a quick,
exploratory analysis and for obtaining starting values for a Bayesian
implementation of the plateau envelope fitted via MCMC.
}

