% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map.plot.R
\name{map.plot}
\alias{map.plot}
\title{Plot maps showing predictions for a given set of climate data.}
\usage{
map.plot(inputs, x.clim.new, x.clim, coordinates, species.name = "",
  scenario.name = "", save.PDF = FALSE, file.name)
}
\arguments{
\item{inputs}{If a numeric vector, then it contains the envelope parameters and has length 2p+p+2+p(p-1)/2.
If a logical vector, then it should have length n. The former is used for plotting shaded predictions for
the climate envelope models; the latter is used for making simple presence/absence plots. For presence/absence
plots, \code{x.clim.new} and \code{x.clim} are ignored.}

\item{x.clim.new}{A matrix, having the same columns as \code{x.clim}, and
containing the points in climate space at which to evaluate the
envelope and produce predictions in space.}

\item{x.clim}{The n by p matrix of climate covariates.}

\item{coordinates}{An n by 2 matrix or list or data frame, having elements
 (or columns) \code{long} and \code{lat} for longitude and latitude
respectively.}

\item{species.name}{The pretty name of the species for plotting.}

\item{scenario.name}{The name of the scenario (and year) for plotting.}

\item{save.PDF}{Logical to indicate whether to save as PDF (default
\code{FALSE}).}

\item{file.name}{If \code{save.PDF} is \code{TRUE}, the name of the pdf
file; if missing, this is created using the species and scenario names
(if available).}
}
\value{
Map of predictions of "climate suitability".
}
\description{
\code{map.plot} plots the fitted maps of "climate suitability" for a given
set of climate variable values. Can also be use to plot maps of the presence/absence data.
}
\seealso{
\code{link{envelope.plot}}
}

