% Generated by roxygen2 (4.0.2): do not edit by hand
\name{map.plot}
\alias{map.plot}
\title{Plot maps showing predictions for a given set of climate data.}
\usage{
map.plot(pars, x.clim.new, x.clim, coordinates, species.name = "",
  scenario.name = "", save.PDF = FALSE, file.name)
}
\arguments{
\item{pars}{The vector of envelope parameters, length 2p+p+2+p(p-1)/2.}

\item{x.clim}{A matrix, the points in climate space at which to evaluate the
envelope and produce predictions in space.}

\item{coordinates}{An n by 2 matrix or list or data frame, having elements
 (or columns) \code{long} and \code{lat} for longitude and latitude
respectively.}

\item{species.name}{The pretty name of the species for plotting.}

\item{scenario.name}{The name of the scenario (and year) for plotting.}

\item{save.PDF}{Logical to indicate whether to save as PDF (default
\code{FALSE}).}

\item{file.name}{If \code{save.PDF} is \code{TRUE}, the name of the pdf
file; if missing, this is created using the species and scenario names
(if available).}
}
\value{
Map of predictions of "climate suitability".
}
\description{
\code{map.plot} plots the fitted maps of "climate suitability" for a given
set of climate variable values.
}
\seealso{
\code{link{envelope.plot}}
}

