% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plateau-package.R
\docType{package}
\name{plateau-package}
\alias{plateau}
\alias{plateau-package}
\title{Fit plateau climate envelopes in species distribution models for presence/absence data}
\description{
Fit plateau climate envelopes in species distribution models for presence/absence data
}
\details{
plateau allows you to fit "plateau" (and unimodal) climate envelopes in
species distribution models for presence/absence data

\tabular{ll}{
Package: \tab plateau\cr
Type: \tab Package\cr
Version: \tab 0.2.0\cr
Date: \tab 2016-06-17\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}

\code{\link{envelope.plot}} plots an envelope
}
\author{
Mark J Brewer \email{Mark.Brewer@bioss.ac.uk}
}
\references{
\url{https://github.com/MarkJBrewer/plateau}
}
\keyword{package}

