% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.bugs.model.R
\name{write.bugs.model}
\alias{write.bugs.model}
\title{Write WinBUGS code to a temporary file}
\usage{
write.bugs.model(n.x.clim, n.x.nonclim = 0)
}
\arguments{
\item{n.x.clim}{The number of climate covariates (n); needed as the WinBUGS
code is slightly different for n=1.}

\item{n.x.nonclim}{The number of non-climate covariates; needed to know
whether or not to include a separate loop for such covariates.}
}
\description{
Write WinBUGS code to a temporary file
}

